practice/production_api/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Validation & Serialization
email-validator==2.1.0
python-dateutil==2.8.2

# Environment & Configuration
python-dotenv==1.0.0

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.0.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# CORS
python-cors==1.0.0

# Database Connection Pool
sqlalchemy[asyncio]==2.0.25

# Redis (for caching/sessions)
redis==5.0.1
aioredis==2.0.1

# Celery (for background tasks)
celery==5.3.4

# Prometheus (for metrics)
prometheus-client==0.19.0

# Sentry (for error tracking)
sentry-sdk==1.39.2
